+++ src/trx/game/inject/editors/objects.c
@@ -30,21 +30,7 @@ static void M_ObjectLinkEdits(
         const INJECTION_OBJECT_INFO source_obj_info =
             Inject_ReadObjectPtr(injection);
 
-        OBJECT *const base_obj = Object_TryGet(base_obj_info.id);
-        const OBJECT *const source_obj = Object_TryGet(source_obj_info.id);
-        if (base_obj == nullptr || source_obj == nullptr
-            || !source_obj->loaded) {
-            continue;
-        }
-
-        base_obj->frame_base = source_obj->frame_base;
-        base_obj->frame_ofs = source_obj->frame_ofs;
-        base_obj->anim_idx = source_obj->anim_idx;
-        base_obj->anim_count = source_obj->anim_count;
-        base_obj->bone_idx = source_obj->bone_idx;
-        base_obj->mesh_idx = source_obj->mesh_idx;
-        base_obj->mesh_count = source_obj->mesh_count;
-        base_obj->loaded = true;
+        IGNORE(Object_BorrowContent(base_obj_info.id, source_obj_info.id));
     }
 }